test(trace): split generic runner smoke coverage
